Make Github node/edge "types" into strings
Previously, we tried to very cleverly make enforce linking {node/edge}
types with their payloads via dependent typing. As part of this
atttempt, we made every node/edge type into a Flow type. Unfortunately,
this did not work out well, and we decided to go with the simpler
approach of making the types string constants, and ensuring that the
correct type is used ourselves.

As a result, the type complexity of the GitHub plugin decreases
significantly.

Test plan:
It removes complexity from the typesystem, without changing the actual
code that executes except for a few variable renames. Observe that unit
tests still pass, and flow is satisfied.